requirements.txt
openai-whisper
deep-translator
torch
ffmpeg-python
gdown
tqdm
